Add much better access to em() method, separate doctrine loading from modelwrapper.
Now the \model\em() function is available for easy usage. Simply use
em()->persist($object); anywhere in your class.

Note that em() is a function that references a singleton class (
DoctrineLoader ). em() is not per application but depends on the super
application currently running ( hence singleton ). We created it like
this to get easy access to entity manager anywhere without needing to
extend classes.