Added XSS escaping for Objects
Before there was no escaping of objects properties that was assigned to
the view, now a clone is made of objects assigned to the view - this
means that all properties is escaped ( arrays , objects, you name it ).
Now it's safe to assign raw objects (custom onces to ) , to the view.

Removed RestController as it's deprecated, it's outdated and not in use,
Response object is the replacement.